checker: `s[i]` yields `u8`, the byte type, not `i32` (#5629) (#5869)
Acceptance item 2 of #5629. A byte and "some integer" were the same type, so
byte-level code type-checked against arbitrary arithmetic and the byte/scalar
distinction survived only as a naming convention. `char` was already distinct;
`u8` was not.

The type-side change is two lines -- both `*ast.Index` string branches
(StringType and the `str` view) now return `NumberType{Width: 8}`. Everything
else is the migration that forces, which is atomic: the tree does not compile
in between.

    var b: u8 = s[0];         // was error[E003]
    var n: i32 = s[0];        // now error[E003]
    var n: i32 = s[0] as i32; // explicit widening

Fern has no implicit unsigned widening, so a byte reaching i32 arithmetic takes
an explicit `as i32`. The cast must land BEFORE any shift: `(s[i] & 63) << 6`
computed in u8 overflows 8 bits and would silently corrupt multi-byte UTF-8
decoding. The checker catches it as E009 rather than miscompiling, but the
decoders in std/string and std/utf8 widen first, and e2e cases pin 2-/3-byte
decodes for exactly that.

Six byte-taking helpers flip at their DEFINITIONS rather than their call sites
-- `__is_ascii_ws`, `__utf8_is_cont`, `__byte_in_set`, `__radix_digit`, `_dig`,
`repeat_char` -- removing ~50 argument mismatches in one edit. std/unicode.fern
is generated, so its share lands in cmd/unicodegen.

No representation change: `u8` occupies the same slot as `i32`.

What this deliberately does NOT do, both learned by reverting an attempt:

* The 21 ascii classifiers keep their `i32` receivers (slice 3' of #5629).
  Moving them to `u8` made std/string IR-ineligible in the self-host -- which
  registers methods by receiver name while tagging u8 as i32 -- so every
  program importing it fell back from the IR path to the AST emitter.
* `u8` gets no Display/Eq/Ord/Hash impls, for the same reason: they collided
  with the i32 impls and the self-host checker reported E021 where the Go
  checker reports nothing. Byte call sites widen instead.

One latent self-host bug surfaced and is fixed here: the SSA evaluator's unary
case treated everything except `-` and the f64 bit-reinterprets as logical NOT,
so `x as i32` (a unary `as_i32` after the parser's desugar) evaluated `!65` = 0.
Harmless while byte reads yielded i32 and needed no cast.

Tests: three checker tests replace the one pinning `s[i]` as i32; a new
internal/e2e/byte_type_test.go pins the runtime half on x86-64 + arm64,
including the shift-precedence trap and 2-/3-byte UTF-8 decodes. Three cases
were deleted rather than converted -- `(0 - 1).is_ascii()`, `(0 - 1).hex_digit()`
and `(256).to_ascii_string()` pinned negative and out-of-range bytes, which u8
makes unrepresentable; coercing `(256)` would have silently become a test of 0.

Verified: 154/154 example programs produce byte-identical output against the
pre-change compiler; all 28 self-host shards, both unit lanes, all e2e and
differential shards green.

+++ cmd/unicodegen/main.go
@@ -1434,7 +1434,7 @@ function _dig(c: i32): i32 {
 
 // _fld reads the 24-bit field at character offset ` + "`i`" + `.
 function _fld(t: string, i: i32): i32 {
-    return (_dig(t[i]) << 18) | (_dig(t[i + 1]) << 12) | (_dig(t[i + 2]) << 6) | _dig(t[i + 3]);
+    return (_dig(t[i] as i32) << 18) | (_dig(t[i + 1] as i32) << 12) | (_dig(t[i + 2] as i32) << 6) | _dig(t[i + 3] as i32);
 }
 
 // _case_apply maps ` + "`cp`" + ` through the case table — uppercase when
@@ -1504,7 +1504,7 @@ function _ascii_fold(s: string, from: i32, to: i32): string {
     var buf: u8[] = __alloc_u8(n);
     var i: i32 = 0;
     while (i < n) {
-        var b: i32 = s[i];
+        var b: i32 = s[i] as i32;
         if (b >= from && b < from + 26) { b = b + (to - from); }
         buf = buf.with(i, b as u8);
         i = i + 1;
@@ -1720,8 +1720,8 @@ pub function eq_ignore_case(a: string, b: string): boolean {
         if (na != b.len()) { return false; }
         var k: i32 = 0;
         while (k < na) {
-            var ca: i32 = a[k];
-            var cb: i32 = b[k];
+            var ca: i32 = a[k] as i32;
+            var cb: i32 = b[k] as i32;
             if (ca >= 65 && ca <= 90) { ca = ca + 32; }
             if (cb >= 65 && cb <= 90) { cb = cb + 32; }
             if (ca != cb) { return false; }
@@ -1750,7 +1750,7 @@ pub function swap_case(s: string): string {
         var buf: u8[] = __alloc_u8(n);
         var k: i32 = 0;
         while (k < n) {
-            var b: i32 = s[k];
+            var b: i32 = s[k] as i32;
             if (b >= 65 && b <= 90) { b = b + 32; }
             else if (b >= 97 && b <= 122) { b = b - 32; }
             buf = buf.with(k, b as u8);
