+++ internal/e2eselfhost/self_host_asm_ir_path_test.go
@@ -479,12 +479,12 @@ func TestSelfHostAsmIRPath(t *testing.T) {
 		// reuses asm.fern's 16-byte `[data@0,len@8]` box + __fern_str_concat/_eq
 		// helpers, so exit codes must match the AST path exactly.
 		{"str-len", `function main(): i32 { var s = "hello"; return s.len(); }`},
-		{"str-index-local", `function main(): i32 { var s = "hello"; return s[0]; }`},
-		{"str-index-loop", `function main(): i32 { var s = "abc"; var sum = 0; var i = 0; while (i < 3) { sum = sum + s[i]; i = i + 1; } return sum % 200; }`},
-		{"str-index-param", `function first(s: string): i32 { return s[0]; } function main(): i32 { return first("Z"); }`},
+		{"str-index-local", `function main(): i32 { var s = "hello"; return s[0] as i32; }`},
+		{"str-index-loop", `function main(): i32 { var s = "abc"; var sum = 0; var i = 0; while (i < 3) { sum = sum + (s[i] as i32); i = i + 1; } return sum % 200; }`},
+		{"str-index-param", `function first(s: string): i32 { return s[0] as i32; } function main(): i32 { return first("Z"); }`},
 		{"str-slice-len", `function main(): i32 { var s = "hello"; var t = s[1:4]; return t.len(); }`},
-		{"str-slice-idx0", `function main(): i32 { var s = "hello"; var t = s[1:4]; return t[0]; }`},
-		{"str-slice-chain", `function main(): i32 { return "hello"[1:4][2]; }`},
+		{"str-slice-idx0", `function main(): i32 { var s = "hello"; var t = s[1:4]; return t[0] as i32; }`},
+		{"str-slice-chain", `function main(): i32 { return "hello"[1:4][2] as i32; }`},
 		{"str-literal-len", `function main(): i32 { return "world!".len(); }`},
 		{"str-empty-len", `function main(): i32 { var s = ""; return s.len(); }`},
 		{"str-concat-len", `function main(): i32 { var a = "ab"; var b = "cde"; var c = a + b; return c.len(); }`},
@@ -1371,16 +1371,16 @@ func TestSelfHostAsmIRPath(t *testing.T) {
 		// / expression / method-call / negation interpolants, adjacent interpolants,
 		// an interpolant-then-literal, the `{{` brace escape, and empty / plain.
 		{"fstring-i32", `function main(): i32 { var n = 7; var s = f"n={n}!"; return s.len(); }`},
-		{"fstring-i32-char", `function main(): i32 { var n = 7; var s = f"n={n}!"; return s[2]; }`},
+		{"fstring-i32-char", `function main(): i32 { var n = 7; var s = f"n={n}!"; return s[2] as i32; }`},
 		{"fstring-str", `function main(): i32 { var w = "xy"; var s = f"[{w}]"; return s.len(); }`},
-		{"fstring-expr", `function main(): i32 { var a = 10; var s = f"v={a * 2}"; return s[2]; }`},
+		{"fstring-expr", `function main(): i32 { var a = 10; var s = f"v={a * 2}"; return s[2] as i32; }`},
 		{"fstring-method", `function main(): i32 { var w = "hi"; return f"v={w.len()}".len(); }`},
 		{"fstring-neg", `function main(): i32 { var a = 5; return f"v={0 - a}".len(); }`},
 		{"fstring-multi", `function main(): i32 { var a = 1; var b = 2; return f"{a}{b}".len(); }`},
 		{"fstring-interp-lit", `function main(): i32 { var n = 42; return f"{n} apples".len(); }`},
 		{"fstring-empty", `function main(): i32 { return f"".len(); }`},
 		{"fstring-plain", `function main(): i32 { return f"plain".len(); }`},
-		{"fstring-esc-brace", `function main(): i32 { var s = f"a{{b"; return s[1]; }`},
+		{"fstring-esc-brace", `function main(): i32 { var s = f"a{{b"; return s[1] as i32; }`},
 		// ASCII case transforms → fresh string (op_str_to_upper / _to_lower). The
 		// AST path emits __fern_str_to_upper/_lower (str_search runtime); the IR
 		// path emits its own emit_ir_str_case bodies — lengths/bytes must match.
@@ -1467,7 +1467,7 @@ func TestSelfHostAsmIRPath(t *testing.T) {
 		// with asm.fern's register-ABI __fern_chr. Covers len, indexed byte, and
 		// `+` concat of two chr results.
 		{"chr-len", `function main(): i32 { return chr(65).len(); }`},
-		{"chr-byte", `function main(): i32 { return chr(122)[0]; }`},
+		{"chr-byte", `function main(): i32 { return chr(122)[0] as i32; }`},
 		{"chr-concat", `function main(): i32 { var s = chr(72) + chr(105); return s.len() * 100 + s[0]; }`},
 		// String chars -> string[] of 1-char strings (op_str_chars; result is_arr +
 		// is_strarr like split). AST emits __fern_str_chars; IR emits emit_ir_str_chars.
@@ -1699,9 +1699,9 @@ func TestSelfHostAsmIRPath(t *testing.T) {
 		// returned byte proves the interpolant text actually landed: `f"n={7}!"`
 		// → "n=7!" (s[2]='7'=55); `f"v={10*2}"` → "v=20" (s[2]='2'=50); the `{{`
 		// escape → "a{b" (s[1]='{'=123).
-		{"fstring-i32-val", `function main(): i32 { var n = 7; var s = f"n={n}!"; return s[2]; }`, 55},
-		{"fstring-expr-val", `function main(): i32 { var a = 10; var s = f"v={a * 2}"; return s[2]; }`, 50},
-		{"fstring-esc-brace-val", `function main(): i32 { var s = f"a{{b"; return s[1]; }`, 123},
+		{"fstring-i32-val", `function main(): i32 { var n = 7; var s = f"n={n}!"; return s[2] as i32; }`, 55},
+		{"fstring-expr-val", `function main(): i32 { var a = 10; var s = f"v={a * 2}"; return s[2] as i32; }`, 50},
+		{"fstring-esc-brace-val", `function main(): i32 { var s = f"a{{b"; return s[1] as i32; }`, 123},
 		{"fstring-len-val", `function main(): i32 { var n = 42; return f"{n} apples".len(); }`, 9},
 		// boolean-element tuple returns, pinned to the absolute IR-path value (the
 		// AST==IR differential is blind to a both-paths-wrong result). The boolean
@@ -1861,7 +1861,7 @@ func TestSelfHostAsmIRPath(t *testing.T) {
 		// Two random_i32 draws differ (a stuck/zero generator returns 0/1).
 		{"random-i32-varies", `function main(): i32 { var a: i32 = random_i32(); var b: i32 = random_i32(); if (a == 0) { return 0; } if (a == b) { return 1; } return 7; }`, 7},
 		// A random byte is in 0..255.
-		{"random-bytes-byte-range", `function main(): i32 { var s: string = random_bytes(4); var x: i32 = s[0]; if (x >= 0) { if (x <= 255) { return 1; } } return 0; }`, 1},
+		{"random-bytes-byte-range", `function main(): i32 { var s: string = random_bytes(4); var x: i32 = s[0] as i32; if (x >= 0) { if (x <= 255) { return 1; } } return 0; }`, 1},
 		// uuid_v4: 36 chars, '4' at index 14, '-' at 8, distinct draws.
 		{"uuid-v4", uuidV4Program, 0},
 		// Range-for through the x86-64 self-host IR path (#2699). The legacy
