+++ internal/e2eselfhost/self_host_headers_ir_test.go
@@ -32,7 +32,7 @@ function lower(s: string): string {
     var out: string = "";
     var i: i32 = 0;
     while (i < s.len()) {
-        var c: i32 = s[i];
+        var c: i32 = s[i] as i32;
         if (c >= 65 && c <= 90) { out = out + alpha[c - 65 : c - 65 + 1]; }
         else { out = out + s[i:i+1]; }
         i = i + 1;
